Use the same `computeRetaggings` method everywhere
